fix(review): move TestWritePidExceptionSwallow to tests/fleet/test_api.py
TestWritePidExceptionSwallow tests _write_pid from fleet/_api.py.
Moving it from test_state.py to a dedicated test_api.py keeps coverage
for _api.py co-located with its source module.
